refactor(analysis): extract AnalysisDetailHeader to stay under 400-line limit
AnalysisDetail.tsx reached 510 lines after the regression fix was added.
Extracted the Header component, COVERAGE_CONTEXT_QUERY_KEYS, and
getDisplaySignature into a new AnalysisDetailHeader.tsx module.

AnalysisDetail.tsx: 510 → 363 lines
AnalysisDetailHeader.tsx: 165 lines (new)

No behaviour changes.
